fix(iot): roll mqtt tests into iot to reuse helpers, fix package size from 16mb to 128kb
